Refactor: Remove Login/Register logic and implement direct Xtream login
- Remove Login/Register pages and related routes
- Create new XtreamLogin component as main entry point
- Modify auth service to handle Xtream login directly
- Update App.tsx to use new authentication flow
- Remove PrivateRoute component
- Update Profile page to remove Xtream credentials section
- Update auth slice to match new Xtream-only flow
