+++ frontend/src/App.tsx
@@ -1,31 +1,35 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
 import { Box } from '@mui/material';
+import { useSelector } from 'react-redux';
 import Layout from './components/Layout';
-import Login from './pages/Login';
-import Register from './pages/Register';
+import XtreamLogin from './components/XtreamLogin';
 import Home from './pages/Home';
 import LiveTV from './pages/LiveTV';
 import Movies from './pages/Movies';
 import Series from './pages/Series';
 import Favorites from './pages/Favorites';
 import Profile from './pages/Profile';
-import PrivateRoute from './components/PrivateRoute';
+import { RootState } from './store';
 
 function App() {
+  const { user } = useSelector((state: RootState) => state.auth);
+
   return (
     <Box sx={{ minHeight: '100vh', bgcolor: 'background.default' }}>
       <Routes>
-        <Route path="/login" element={<Login />} />
-        <Route path="/register" element={<Register />} />
-        <Route path="/" element={<PrivateRoute><Layout /></PrivateRoute>}>
-          <Route index element={<Home />} />
-          <Route path="live" element={<LiveTV />} />
-          <Route path="movies" element={<Movies />} />
-          <Route path="series" element={<Series />} />
-          <Route path="favorites" element={<Favorites />} />
-          <Route path="profile" element={<Profile />} />
-          <Route path="*" element={<Navigate to="/" replace />} />
-        </Route>
+        {!user ? (
+          <Route path="*" element={<XtreamLogin />} />
+        ) : (
+          <Route path="/" element={<Layout />}>
+            <Route index element={<Home />} />
+            <Route path="live" element={<LiveTV />} />
+            <Route path="movies" element={<Movies />} />
+            <Route path="series" element={<Series />} />
+            <Route path="favorites" element={<Favorites />} />
+            <Route path="profile" element={<Profile />} />
+            <Route path="*" element={<Navigate to="/" replace />} />
+          </Route>
+        )}
       </Routes>
     </Box>
   );
