--- frontend/src/components/PrivateRoute.tsx
@@ -1,15 +1,0 @@
-import { Navigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import { RootState } from '../store';
-
-interface PrivateRouteProps {
-  children: React.ReactNode;
-}
-
-const PrivateRoute: React.FC<PrivateRouteProps> = ({ children }) => {
-  const { isAuthenticated } = useSelector((state: RootState) => state.auth);
-
-  return isAuthenticated ? <>{children}</> : <Navigate to="/login" />;
-};
-
-export default PrivateRoute;
