+++ frontend/src/pages/Profile.tsx
@@ -27,33 +27,8 @@ const Profile = () => {
   const dispatch = useDispatch();
   const { user } = useSelector((state: RootState) => state.auth);
   const { watchHistory } = useSelector((state: RootState) => state.content);
-  const [xtreamCredentials, setXtreamCredentials] = useState({
-    xtream_username: user?.xtream_username || '',
-    xtream_password: user?.xtream_password || '',
-    xtream_url: user?.xtream_url || '',
-  });
   const [error, setError] = useState('');
   const [success, setSuccess] = useState('');
-
-  const handleCredentialsChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setXtreamCredentials({
-      ...xtreamCredentials,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const handleCredentialsSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    try {
-      const response = await auth.updateXtreamCredentials(xtreamCredentials);
-      dispatch(setUser(response.data.user));
-      setSuccess('Xtream credentials updated successfully');
-      setError('');
-    } catch (error: any) {
-      setError(error.response?.data?.error || 'Error updating credentials');
-      setSuccess('');
-    }
-  };
 
   const handleAdultContentToggle = async () => {
     try {
@@ -87,61 +62,7 @@ const Profile = () => {
       </Typography>
 
       <Grid container spacing={3}>
-        <Grid item xs={12} md={6}>
-          <Paper sx={{ p: 3 }}>
-            <Typography variant="h6" gutterBottom>
-              Xtream Credentials
-            </Typography>
-            {error && (
-              <Alert severity="error" sx={{ mb: 2 }}>
-                {error}
-              </Alert>
-            )}
-            {success && (
-              <Alert severity="success" sx={{ mb: 2 }}>
-                {success}
-              </Alert>
-            )}
-            <form onSubmit={handleCredentialsSubmit}>
-              <TextField
-                fullWidth
-                margin="normal"
-                label="Xtream Username"
-                name="xtream_username"
-                value={xtreamCredentials.xtream_username}
-                onChange={handleCredentialsChange}
-              />
-              <TextField
-                fullWidth
-                margin="normal"
-                label="Xtream Password"
-                name="xtream_password"
-                type="password"
-                value={xtreamCredentials.xtream_password}
-                onChange={handleCredentialsChange}
-              />
-              <TextField
-                fullWidth
-                margin="normal"
-                label="Xtream URL"
-                name="xtream_url"
-                value={xtreamCredentials.xtream_url}
-                onChange={handleCredentialsChange}
-                placeholder="http://example.com:port"
-              />
-              <Button
-                type="submit"
-                variant="contained"
-                color="primary"
-                sx={{ mt: 2 }}
-              >
-                Update Credentials
-              </Button>
-            </form>
-          </Paper>
-        </Grid>
-
-        <Grid item xs={12} md={6}>
+        <Grid item xs={12}>
           <Paper sx={{ p: 3 }}>
             <Typography variant="h6" gutterBottom>
               Preferences
